refactor: rename ShortcutsDb -> Shortcuts
